@@ -25,13 +25,13 @@ namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Tests
             var mockTriggeredFunctionExecutor = new Mock<ITriggeredFunctionExecutor>();
 
             mockMqttConnection
-                .Setup(x => x.SubscribeAsync(It.IsAny<TopicFilter[]>()))
+                .Setup(x => x.SubscribeAsync(It.IsAny<MqttTopicFilter[]>()))
                 .Returns(Task.CompletedTask);
 
             mockTriggeredFunctionExecutor
                 .Setup(x => x.TryExecuteAsync(It.IsAny<TriggeredFunctionData>(), It.IsAny<CancellationToken>()));
 
-            var topicFilter = new TopicFilter[] { new TopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
+            var topicFilter = new MqttTopicFilter[] { new MqttTopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
             var mqttListener = new MqttListener(mockMqttConnection.Object, topicFilter, mockTriggeredFunctionExecutor.Object, _mockLogger.Object);
 
             // Act
@@ -51,14 +51,14 @@ namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Tests
             var mockTriggeredFunctionExecutor = new Mock<ITriggeredFunctionExecutor>();
 
             mockMqttConnection
-                .Setup(x => x.SubscribeAsync(It.IsAny<TopicFilter[]>()))
+                .Setup(x => x.SubscribeAsync(It.IsAny<MqttTopicFilter[]>()))
                 .Returns(Task.CompletedTask);
 
             mockMqttConnection
                 .Setup(x => x.UnubscribeAsync(It.IsAny<string[]>()))
                 .Returns(Task.CompletedTask);
 
-            var topicFilter = new TopicFilter[] { new TopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
+            var topicFilter = new MqttTopicFilter[] { new MqttTopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
             var mqttListener = new MqttListener(mockMqttConnection.Object, topicFilter, mockTriggeredFunctionExecutor.Object, _mockLogger.Object);
 
             // Act
@@ -77,13 +77,13 @@ namespace CaseOnline.Azure.WebJobs.Extensions.Mqtt.Tests
             var mockTriggeredFunctionExecutor = new Mock<ITriggeredFunctionExecutor>();
 
             mockMqttConnection
-                .Setup(x => x.SubscribeAsync(It.IsAny<TopicFilter[]>()))
+                .Setup(x => x.SubscribeAsync(It.IsAny<MqttTopicFilter[]>()))
                 .Returns(Task.CompletedTask);
 
             mockTriggeredFunctionExecutor
                 .Setup(x => x.TryExecuteAsync(It.IsAny<TriggeredFunctionData>(), It.IsAny<CancellationToken>()));
 
-            var topicFilter = new TopicFilter[] { new TopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
+            var topicFilter = new MqttTopicFilter[] { new MqttTopicFilter() { Topic = "test/topic", QualityOfServiceLevel = MQTTnet.Protocol.MqttQualityOfServiceLevel.AtLeastOnce } };
             var mqttListener = new MqttListener(mockMqttConnection.Object, topicFilter, mockTriggeredFunctionExecutor.Object, _mockLogger.Object);
 
             // Act
